+++ src/lib/app-provider/hooks/useBaseApiRoute.ts
@@ -3,6 +3,7 @@ import { CELATONE_API_OVERRIDE as api } from "env";
 
 export const useBaseApiRoute = (
   type:
+    | "overviews"
     | "txs"
     | "balances"
     | "assets"
@@ -30,6 +31,8 @@ export const useBaseApiRoute = (
     );
 
   switch (type) {
+    case "overviews":
+      return `${api}/v1/${chain}/${currentChainId}/overviews`;
     case "txs":
       return `${api}/v1/${chain}/${currentChainId}/txs`;
     case "balances":